==> notebooks/01_cleaning_monthly_daily_EPA_2019.ipynb <==
import pandas as pd
import numpy as np

pd.set_option('display.max_columns', None)

raw = pd.read_csv('../data/epa_daily_88101_2019.csv')
df = raw.copy()
df.head()

# DateTime

df['Date Local'] = pd.to_datetime(df['Date Local'])
df.set_index('Date Local', inplace = True)
df.sort_index(inplace = True)

df = df.groupby(['State Code', 'County Code', 'Site Num', 'Address', 
            'State Name', 'County Name', 'City Name', 'Latitude', 'Longitude', 
            'Parameter Name', 'Sample Duration', 'Pollutant Standard'])['Arithmetic Mean'].resample('M').mean().reset_index()

df.rename(columns = {'Arithmetic Mean': 'PM2.5 Monthly Mean'}, inplace = True)

df.head()

df.set_index('Date Local', inplace = True)
df.sort_index(inplace = True)

df.head()

# Clean Lat Lon

# https://science.blurtit.com/255139/what-is-the-longitude-and-latitude-of-the-united-states
us_lat_min = 24.5
us_lat_max = 49.489998

us_lon_min = -124.75
us_lon_max = -66.76

df = df[(df['Latitude'] >= us_lat_min) & (df['Latitude'] <= us_lat_max)]
df = df[(df['Longitude'] >= us_lon_min) & (df['Longitude'] <= us_lon_max)]

df.head()

df.sort_values(by = 'Longitude', ascending = True).tail()

df.shape

# Save CSVs

# df.to_csv('../data/epa_monthly_addresses.csv', index = True)

# df = df[['Latitude', 'Longitude', 'PM2.5 Monthly Mean']]

# df.to_csv('../data/epa_monthly.csv', index = True)
