==> notebooks/02_1_random_lat_lon.ipynb <==
## Generate random latitude and longitude points across the contiguous Unites States

import random
from global_land_mask import globe
import numpy as np
import pandas as pd

Define the maximum and minimum latitude and longitude points in the contiguous US (CONUS): 

north = 49.3
south = 25.4
east = -65.9
west = -124.7

Create a random list of 1200 latitude and longitude points across the CONUS, making sure that the coordinates fall on land and not in the ocean:

count = 0
latlonglist = []
while count < 1200:
    lat = round(random.uniform(south, north), 5)
    lon = round(random.uniform(east, west), 5)
    is_on_land = globe.is_land(lat, lon)
    if is_on_land == True:
        count += 1
        latlonglist.append([lat, lon])

Create a DataFrame and save to CSV file for use in modeling PM2.5 predictions.

df = pd.DataFrame(latlonglist, columns=["lat", "lon"])
# df.to_csv('../data/monthly_test_data/latlong_1k.csv', index = False)
