==> notebooks/04_more_model_testing.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.pipeline import Pipeline

from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.preprocessing import PolynomialFeatures, StandardScaler
from sklearn.neighbors import KNeighborsRegressor

from sklearn.linear_model import LinearRegression, LassoCV
from sklearn import metrics
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import Ridge, Lasso, ElasticNet
from sklearn.naive_bayes import MultinomialNB
from sklearn.ensemble import GradientBoostingRegressor, AdaBoostRegressor, VotingRegressor

from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor, AdaBoostClassifier, AdaBoostRegressor, BaggingClassifier, BaggingRegressor, StackingRegressor

df = pd.read_csv('../data/epa_final.csv')

df.head()

X = df.drop(columns = ['Date Local','PM2.5 Monthly Mean', 'lat_2dec', 'lon_2dec'])
y = df['PM2.5 Monthly Mean']

X.head()

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42)

ss = StandardScaler()
X_train_sc = ss.fit_transform(X_train)
X_test_sc = ss.transform(X_test)

lr = LinearRegression()
lr.fit(X_train_sc, y_train)
lr_train_pred = lr.predict(X_train_sc)
lr_test_pred = lr.predict(X_test_sc)
print(f'LR training score: {lr.score(X_train_sc,y_train)}')
print(f'LR test score: {lr.score(X_test_sc,y_test)}')

# Decision tree model:
dt = DecisionTreeRegressor()
dt.fit(X_train_sc, y_train)
dt_train_pred = dt.predict(X_train_sc)
dt_test_pred = dt.predict(X_test_sc)
print(f'Decision Tree training score: {dt.score(X_train_sc,y_train)}')
print(f'Decision Tree test score: {dt.score(X_test_sc,y_test)}')

bag = BaggingRegressor()
bag.fit(X_train_sc, y_train)
bag_train_pred = bag.predict(X_train_sc)
bag_test_pred = bag.predict(X_test_sc)
print(f'Bagged DT training score: {bag.score(X_train_sc,y_train)}')
print(f'Bagged DT test score: {bag.score(X_test_sc,y_test)}')

rf = RandomForestRegressor()
rf.fit(X_train_sc, y_train)
rf_train_pred = rf.predict(X_train_sc)
rf_test_pred = rf.predict(X_test_sc)
print(f'Random forest training score: {rf.score(X_train_sc,y_train)}')
print(f'Random forest test score: {rf.score(X_test_sc,y_test)}')

ada = AdaBoostRegressor(base_estimator=RandomForestRegressor())
ada.fit(X_train_sc, y_train)
ada_train_pred = ada.predict(X_train_sc)
ada_test_pred = ada.predict(X_test_sc)
print(f'Adaboost training score: {ada.score(X_train_sc,y_train)}')
print(f'Adaboost test score: {ada.score(X_test_sc,y_test)}')

X.shape

# Create the model
level1_models = [
    ('bag', BaggingRegressor()),
    ('knn_pipe', Pipeline([
        ('ss', StandardScaler()),
        ('knn', KNeighborsRegressor())
    ])),
    ('lasso_pipe', Pipeline([
        ('ss', StandardScaler()),
        ('lasso', LassoCV())
    ]))
]

stack = StackingRegressor(estimators=level1_models,
                         final_estimator=RandomForestRegressor())

# Fit
stack.fit(X_train_sc,y_train)

# Train score
stack.score(X_train_sc, y_train)

# Test score
stack.score(X_test_sc, y_test)

level2_models = [
    ('bag', BaggingRegressor()),
    ('ada_pipe', Pipeline([
        ('ss', StandardScaler()),
        ('ada', AdaBoostRegressor(base_estimator=RandomForestRegressor()))
    ])),
    ('lasso_pipe', Pipeline([
        ('ss', StandardScaler()),
        ('lasso', LassoCV())
    ]))
]

stack1 = StackingRegressor(estimators=level2_models,
                         final_estimator=RandomForestRegressor())

# Fit
stack1.fit(X_train_sc,y_train)

# Train score
stack1.score(X_train_sc, y_train)

# Test score
stack1.score(X_test_sc, y_test)

level3_models = [
    ('bag', BaggingRegressor()),
    ('lasso_pipe', Pipeline([
        ('ss', StandardScaler()),
        ('lasso', LassoCV())
    ]))
]

stack2 = StackingRegressor(estimators=level3_models,
                         final_estimator=RandomForestRegressor())

# Fit
stack2.fit(X_train_sc,y_train)

# Train score
stack2.score(X_train_sc, y_train)

# Test score
stack2.score(X_test_sc, y_test)

level4_models = [
    ('bag', BaggingRegressor()),
    ('dt_pipe', Pipeline([
        ('ss', StandardScaler()),
        ('rf', RandomForestRegressor())
    ])),
    ('lasso_pipe', Pipeline([
        ('ss', StandardScaler()),
        ('lasso', LassoCV())
    ]))
]

stack3 = StackingRegressor(estimators=level4_models)

# Fit
stack3.fit(X_train_sc,y_train)

# Train score
stack3.score(X_train_sc, y_train)

# Test score
stack3.score(X_test_sc, y_test)

level5_models = [
    ('bag', BaggingRegressor()),
    ('rf_pipe', Pipeline([
        ('ss', StandardScaler()),
        ('rf', RandomForestRegressor())
    ])),
    ('lasso_pipe', Pipeline([
        ('ss', StandardScaler()),
        ('lasso', LassoCV())
    ]))
]

stack4 = StackingRegressor(estimators=level5_models)

params = {
    'rf_pipe__rf__n_estimators': [92],
    'lasso_pipe__lasso__n_alphas': [105],
    'lasso_pipe__lasso__max_iter': [925],
    'bag__n_estimators': [10],
    'rf_pipe__rf__min_samples_split': [5]
    
}

gs = GridSearchCV(estimator = stack4, param_grid = params, cv = 5)

gs.fit(X_train_sc,y_train)

gs.best_score_

gs.best_params_

gs.score(X_train_sc, y_train)

gs.score(X_test_sc,y_test)

stack4.get_params().keys()

bag.get_params()

rf.get_params()

LassoCV().get_params()
